BUGFIX: now the correct label is used

+++ lagrangeRL/network/lagrangeEligTfApprox.py
@@ -15,7 +15,7 @@ class lagrangeEligTfApprox(lagrangeEligTf):
         Initialize an empty class instance
         """
 
-        self.dtype = tf.float64
+        self.dtype = tf.float32
         self.T = 0.
 
         # set up a logger
